Refactor UI migration structure with compatibility wrappers
