+++ docs/backlog.md
@@ -1,3 +1,4 @@
+- [x] UI-16 Migration UI modulaire (phase 1): création des sous-répertoires `screens/command_center|command_deck|facility|battle_hud`, `components/cards|lists|agent|shared`, `layouts/grid|split|overlays`, `navigation/focus|input`, `feedback/toast|dialog|banner`; extraction initiale `command_deck` vers `game/ui/screens/command_deck/layout.py`; wrappers de compatibilité maintenus; mutualisation badge upgrade + rendu cartes agents via composants dédiés. (completed May 22, 2026)
 - [x] UI-15 Foundation design-system split: création `theme/spacing.py`, `theme/elevation.py`, `theme/radii.py`, normalisation des couleurs sémantiques, ajout de `game/ui/components/foundation/` (Panel/Button/Badge/Divider/ProgressBar/Tooltip), règle progressive anti-hardcode sur `game/ui/screens/`, docs design-system et tests dédiés. (completed May 22, 2026)
 - [x] UI-14 Arborescence écrans UI incrémentale: création de `game/ui/screens/`, déplacement progressif des modules de vue (`command_center`, `dashboard`, `facility`, `mission_board`, `research_lab`) avec wrappers de compatibilité temporaires dans `game/ui/`, et test de non-régression d'imports `tests/ui/test_screen_wrappers.py`. (completed May 22, 2026)
 - [x] UI-11 Feedback system modulaire: création de `game/ui/feedback/` (`toast_center.py`, `confirm_dialog.py`, `error_banner.py`), standardisation des messages `Action/Result/Impact/Next`, confirmations pour actions coûteuses/irréversibles (mission risquée, dépenses, load), et tests de rendu/dispatch `tests/ui/test_feedback_system.py`.
