+++ game/ui/components/cards/__init__.py
@@ -1,0 +1,1 @@
+"""UI package."""
