+++ game/ui/components/shared/__init__.py
@@ -1,0 +1,1 @@
+"""UI package."""
