+++ game/ui/feedback/banner/__init__.py
@@ -1,0 +1,1 @@
+"""UI package."""
