+++ game/ui/feedback/toast/__init__.py
@@ -1,0 +1,1 @@
+"""UI package."""
