+++ game/ui/layouts/split/__init__.py
@@ -1,0 +1,1 @@
+"""UI package."""
