+++ game/ui/navigation/focus/__init__.py
@@ -1,0 +1,1 @@
+"""UI package."""
