+++ game/ui/navigation/input/__init__.py
@@ -1,0 +1,1 @@
+"""UI package."""
