+++ game/ui/screens/command_deck/__init__.py
@@ -1,0 +1,3 @@
+"""Command-deck screen package with small mono-responsibility modules."""
+
+from .layout import *  # noqa: F401,F403
